Temporarily stop using bmt, because it is flaky
bmt makes unnecessary and unreliable use of the network for remote
schema self-validation every time a toolkit is constructed.  Whenever
that remote schema file is unavailable (such as due to SSL certificate
issues), this causes kgx to break.  That seems really silly.  A better
design would keep the schema file local to the bmt installation to avoid
unnecessary reliance on the network.

We currently don't need bmt on this branch, so until bmt is improved, we
can just stop importing modules that depend on it.